Fixing the ec2_instance var to add to new host
In case we terminate the instance or create the instance from scratch, the code
to generate ec2_instances return empty list thus we do not have the data in and
caused error.

The previous code always set the var ec2_instance already in all cases, thus just
re-use that data.

Also output the debugging information for it.

This is the regression of recent changes.