Make route53 records
Many times we need a route53 points to the instance thus implement it in here.

Add two variables ec2_instance_public_dns and ec2_instance_private_dns
If they are defined as string, then create route53 for it.
For public one hardcoded the zone using xvt.technology for now. It wont be
complex to create public zone per aws account but there is no need for it at
the moment.