Move the attributes to the task in include_tasks
The ansibe 2.6.x now completely not passing these attributes in 'include_tasks'
into the task file thus the delegate_to does not take effect and end result is
we try to do it on the current host which does not have the device /dev/xvdf to
setup.

Not seeing this change is smart about ansible - rather stupid but well I have
no control over them.

This to fix the issues.