do-not-use-ec2_eip-when-public-address-is-available
We support a string 'auto' to tell the system to auto allocate eip. Thus if we
depends on the value of ec2_eip we got 'auto' which does not resolve to any
rela host or address at all.

But at this stepo we always have the ec2 object in aws which has the public
address, why not use this as it refects exactly the current host status.