Allow ec2 have its own settings for python interpreter.
There is issues with ansible dnf module using python3. Ideally we need the
python2 in the remote host (centos7)

However when we launch we use default python interpreter for the localhost
which is python3. We can override it using python2 but then we need to install
botto, botto3 for localhost to be able to launch role ec2_instance.

This change allow us at localhost when we launch the ec2 we can use python3 but
in the remote host when the ec2 is launched we can use python set in the
variable ec2_python_interpreter - which can be python2 or 3 or whatever it
suits.